move: component/ui/utils as component/utils
