+++ src/components/pages/FirstProfile/index.tsx
@@ -1,6 +1,6 @@
 import type { NextPageWithLayout } from '@/pages/_app'
 import BrowserLayout from '@/components/layout/BrowserLayout'
-import ErrorDisplay from '@/components/ui/utils/ErrorDisplay';
+import ErrorDisplay from '@/components/utils/ErrorDisplay';
 import FirstProfileForm from './FirstProfileForm';
 import { FirstProfileShellCss } from './styles/FirstProfilePageStyle';
