+++ src/components/pages/Payment/RegistCardFormPage/index.tsx
@@ -2,7 +2,7 @@ import type { NextPageWithLayout } from '@/pages/_app';
 import { useTranslation } from 'next-i18next';
 import SignupLayout from '@/components/layout/SignupLayout';
 import PageInOutAnimate from '@/components/ui/animation/PageInOutAnimate';
-import ConditionalRender from '@/components/ui/utils/ConditionalRender';
+import ConditionalRender from '@/components/utils/ConditionalRender';
 import StepHeader from '../../Signup/component/StepHeader';
 import CardInfoForm from './component/CardInfoForm';
 import CardInfoFormWithPolicys from './component/CardInfoFormWithPolicys';
