+++ src/components/ui/Form/DarkSelect/index.tsx
@@ -1,8 +1,8 @@
 import type { InputLayoutValues } from '../InputLayout';
 import type { ReactNode, SelectHTMLAttributes } from 'react';
 import { useMemo } from 'react';
+import ConditionalRender from '@/components/utils/ConditionalRender';
 import SelectArrow from '@assets/netflix/select-arrow.svg'
-import ConditionalRender from '../../utils/ConditionalRender';
 import InputLayout from '../InputLayout';
 import { InputErrorStateCss, InputThemeCss, InputValidatedStateCss } from '../TextInput/style';
 import { inputCss, layoutCss, outlineCss, postfixCss, prefixCss, shellCss } from './style';
