+++ src/components/utils/Intersection/lib.ts
